updated the returned logic
